pelbot/pyproject.toml
[tool.poetry]
name = "pelbot"
version = "0.1.0"
description = "Louisiana Election Data Scraper"
authors = ["Bentley Hensel <bentleyhensel@gmail.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
